♻️ [refactor] EmotionLottie 로 통일
